Provide support for running the app as container (#35)
* (#2) Add flake8 into list of dev dependencies

* (#2) Reorder application instantiation statements to avoid ImportError

* (#2) Add MONGO_URI into Config object

This commit adds `MONGO_URI` into `Config` object. The value
is constructed from given username and password. The reason
for adding a new property in Config object is the app
refused to start due to lack of URI information to a running
MongoDB instance.

* (#2) Ignore cloned Git repositories

* Put document analysers to separate loop

This commit separates document analysers from count-based
analysers due to difference in `calculate_score()` signature.

* (#2) Add Dockerfile and docker-compose-yml

* (#2) Replace CMD with ENTRYPOINT in Dockerfile

* (#2) Specify DB name in MONGO_URI

* (#2) Fix package name capitalisation in README

* (#2) Update README

* (#2) Bind app port to port 5000

* (#2) Shorten docker-compose call in README

* (#2) Use echo instead of here-string

* Minor grammar fix in README

* Update README